+++ test/features/sync/matrix/pipeline/matrix_stream_consumer_test.dart
@@ -16,6 +16,7 @@ import 'package:lotti/features/sync/matrix/sync_event_processor.dart';
 import 'package:lotti/features/sync/matrix/sync_room_manager.dart';
 import 'package:lotti/services/logging_service.dart';
 import 'package:matrix/matrix.dart';
+import 'package:matrix/src/utils/cached_stream_controller.dart' as matrix_utils;
 import 'package:mocktail/mocktail.dart';
 
 import '../../../../mocks/mocks.dart';
@@ -26,13 +27,23 @@ class MockSyncRoomManager extends Mock implements SyncRoomManager {}
 
 class MockSyncReadMarkerService extends Mock implements SyncReadMarkerService {}
 
-class MockClient extends Mock implements Client {}
+class MockClient extends Mock implements Client {
+  MockClient() {
+    when(() => onSync).thenReturn(
+      matrix_utils.CachedStreamController<SyncUpdate>(),
+    );
+  }
+}
 
 class MockRoom extends Mock implements Room {}
 
 class MockTimeline extends Mock implements Timeline {}
 
-class MockEvent extends Mock implements Event {}
+class MockEvent extends Mock implements Event {
+  MockEvent() {
+    when(() => originServerTs).thenReturn(DateTime(2026, 4, 20));
+  }
+}
 
 class _FakeClient extends Fake implements Client {}